27th commit

+++ projects/research/gsp/graph_signal_restoration/data_handling/graph/create_graph.m
@@ -1,6 +1,6 @@
 function graph = create_graph(weights)
 
-    graph = gsp_graph(double(weights));
+    graph = gsp_graph(weights);
     graph = gsp_compute_fourier_basis(graph);
     graph = gsp_adj2vec(graph);
